backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
drf-spectacular
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
mysqlclient
pandas
scikit-learn
joblib
requests
ulid-py